Move system auth routine to OpenSim.Framework.Util. Non Windows version aleways returns true for now.  Change references to use new location.  Add required refs into prebuild.xml and map tabs to space.
