+++ .gitignore
@@ -6,6 +6,8 @@
 *.user
 *.userosscache
 *.sln.docstates
+*.csproj
+*.sln
 
 # User-specific files (MonoDevelop/Xamarin Studio)
 *.userprefs
