+++ OpenSim/Framework/Util.cs
@@ -53,6 +53,11 @@ using Amib.Threading;
 using System.Drawing;
 using System.Runtime.InteropServices;
 
+#if !__MonoCS__
+using System.DirectoryServices.AccountManagement;
+#endif
+
+
 namespace OpenSim.Framework
 {
     /// <summary>
@@ -168,6 +173,32 @@ namespace OpenSim.Framework
                 return GetTickCount64();
             #endif 
         }
+
+
+		/// <summary
+		/// Authenticate a username/password pair against the user we are running under.
+		/// </summary>
+		/// <remarks>checks that the username is the same as the current System.Environment.UserName,
+		/// And Validates the password against the password for that account</remarks>
+		/// <returns>true if the authentication succeeded, false otherwise.</returns>
+		/// <param name='username'>string</param>
+		/// <param name='password'>string</param>
+		/// 
+		public static bool AuthenicateAsSystemUser(string username, string password)
+		{
+#if __MonoCS__
+			return true;
+#else
+			// Is the username the same as the logged in user and do they have the password correct?
+			PrincipalContext pc = new PrincipalContext(ContextType.Machine);
+			bool isValid = 
+			(username.Equals(System.Environment.UserName) && 
+			pc.ValidateCredentials(username, password));
+
+			return (isValid);
+#endif
+		}
+
 
         #region Vector Equations
